Free me from this purgatory

+++ src/main/java/com/Da_Technomancer/crossroads/items/crafting/recipes/IceboxRec.java
@@ -31,7 +31,7 @@ public class IceboxRec implements IRecipe<IInventory>{
 		this.cooling = (float) cooling;
 	}
 
-	public double getCooling(){
+	public float getCooling(){
 		return cooling;
 	}
